Move aligned_producer test to its own file
